refactor(python-sort): Refactor and add Python sorting algorithms to proper folders
