+++ python/algorithms/sorting/insertion_sort/insertion_sort.py
