+++ python/algorithms/sorting/selection_sort/selection_sort.py
