+++ python/tests/algorithms/sorting/test_insertion_sort.py
@@ -1,5 +1,5 @@
 import pytest
-from algorithms.sorting.insertion_sort import insertion_sort
+from python.algorithms.sorting.insertion_sort.insertion_sort import insertion_sort
 
 def test_insertion_sort_empty():
     """
